Refactor win checking
The win checking function previously return a bool for whether
the game is won or not. The tie checking was done at the `Game`
level. This made it difficult to unit test the tie checking.

This commit refactors the win checking to return a board state,
which would indicate whether the game is won, tied, or being played.
This allows it to be unit tested much more easily.