Initial board metadata implementation
Implements a metadata structure for the board. This structure
finds and keeps track of winning moves for a specific marker.
It also keeps a vector of reactors (functions) for each cell.
This way, we can trigger other actions and updates to the metadata
when a player marks a specific cell.