Allow human player to select computer opponent type.
Now that the human user selects their opponent, we don't
know in advance which players will be needed to create the
Game structure. In this case, we use trait objects to take
advantage of dynamic dispatch.

I had to create a Player trait and have all the different
player structs implement that trait. That way, we can Box
a concret player object and place it in the Game.

Currently, resetting the Game is not implemented correctly.
Resetting the game should restart it with the same player type.
More on this in future commits.